feat(model): 分离产品调拨单和历史记录模型
